Test.HUnit.Terminal

Description

This module handles the complexities of writing information to the terminal, including modifying text in place.

Documentation

terminalAppearance :: String -> String

Simplifies the input string by interpreting \r and \b characters specially so that the result string has the same final (or terminal, pun intended) appearance as would the input string when written to a terminal that overwrites character positions following carriage returns and backspaces.