Test.HUnit.Lang

Documentation

type Assertion = IO ()

When an assertion is evaluated, it will output a message if and only if the assertion fails.

Test cases are composed of a sequence of one or more assertions.

assertFailure

Arguments

:: (?loc :: CallStack)

=> String
A message that is displayed with the assertion failure

-> Assertion

Unconditionally signals that a failure has occured. All other assertions can be expressed with the form:

   if conditionIsMet
       then IO ()
       else assertFailure msg

data Location

Constructors

Location

Fields

locationFile :: FilePath

locationLine :: Int

locationColumn :: Int

data Result

Constructors

Success

Failure (Maybe Location) String

Error (Maybe Location) String

performTestCase

Arguments

:: Assertion
an assertion to be made during the test case run

-> IO Result

Performs a single test case.

Internals

Note: This is not part of the public API! It is exposed so that you can tinker with the internals of HUnit, but do not expect it to be stable!

data HUnitFailure

Constructors

HUnitFailure (Maybe Location) String