Test.HUnit.Lang

Documentation

type Assertion = IO ()

When an assertion is evaluated, it will output a message if and only if the assertion fails.

Test cases are composed of a sequence of one or more assertions.

assertFailure

Arguments

:: HasCallStack

=> String
A message that is displayed with the assertion failure

-> IO a

Unconditionally signals that a failure has occured. All other assertions can be expressed with the form:

   if conditionIsMet
       then IO ()
       else assertFailure msg

assertEqual

Arguments

:: (HasCallStack, Eq a, Show a)

=> String
The message prefix

-> a
The expected value

-> a
The actual value

-> Assertion

Asserts that the specified actual value is equal to the expected value. The output message will contain the prefix, the expected value, and the actual value.

If the prefix is the empty string (i.e., ""), then the prefix is omitted and only the expected and actual values are output.

data Result

Constructors

Success

Failure (Maybe SrcLoc) String

Error (Maybe SrcLoc) String

performTestCase

Arguments

:: Assertion
an assertion to be made during the test case run

-> IO Result

Performs a single test case.

Internals

Note: This is not part of the public API! It is exposed so that you can tinker with the internals of HUnit, but do not expect it to be stable!

data HUnitFailure

Constructors

HUnitFailure (Maybe SrcLoc) FailureReason

data FailureReason

Constructors

Reason String

ExpectedButGot (Maybe String) String String

formatFailureReason :: FailureReason -> String